Last cleanup int -> intptr_t. Also removed a hack (boolean is not an integer).
R=asiva@google.com

Review URL: https://codereview.chromium.org//26345002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@28372 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/assembler_arm.cc
@@ -1819,7 +1819,7 @@ void Assembler::EmitBranch(Condition cond, Label* label, bool link) {
       EmitType5(cond, dest, link);
     }
   } else {
-    const int position = buffer_.Size();
+    const intptr_t position = buffer_.Size();
     if (use_far_branches()) {
       const int32_t dest = label->position_;
       EmitFarBranch(cond, dest, link);
